requirements.txt
openai>=1.45.0
pydantic>=2.9.0
python-dotenv>=1.0.1
numpy>=1.26.4
pydub>=0.25.1
python-docx>=1.1.0
pypdf>=5.1.0
